+++ include/inc_lib/checkmessage.inc.php
@@ -21,7 +21,7 @@ $sql = "SELECT COUNT(*) FROM ".DB_PREPEND."phpwcms_message WHERE msg_uid=".intva
 if(($check = _dbQuery($sql, 'COUNT'))) {
 
     $wcsnav["navspace1"]  = '<a href="phpwcms.php?do=messages" title="'.$check.' new messages waiting!">'.
-    $wcsnav["navspace1"] .= '<img src="img/symbole/new_mail.gif" border="0" alt="" /></a>';
+    $wcsnav["navspace1"] .= '<i class="fas fa-envelope text-warning ml-1"></i></a>';
     $new_mail_waiting = 1;
 
 }