+++ include/inc_lib/news.inc.php
@@ -200,23 +200,23 @@ class phpwcmsNews {
                 $next_page = $max_page - 1;
             }
 
-            $paginate .= '<img src="img/famfamfam/action_back.gif" alt="" border="0"';
+            $paginate .= '<i class="fas fa-chevron-left text-primary cursor-pointer';
             if($current_page == 1) {
-                $paginate .= ' class="inactive"';
+                $paginate .= ' text-muted inactive"';
             } else {
-                $paginate .= ' onclick="$(\'filterPage\').value='.$prev_page.';$(\'paginate\').submit();"';
+                $paginate .= '" onclick="$(\'filterPage\').value='.$prev_page.';$(\'paginate\').submit();"';
             }
-            $paginate .= ' /></td><td class="chatlist text-nowrap">';
+            $paginate .= '></i></td><td class="chatlist text-nowrap">';
 
             $paginate .= '&nbsp;<b>' . $current_page . '</b>/' . $max_page . '&nbsp;';
 
-            $paginate .= '</td><td class="text-nowrap"><img src="img/famfamfam/action_forward.gif" alt="" border="0"';
+            $paginate .= '</td><td class="text-nowrap"><i class="fas fa-chevron-right text-primary cursor-pointer';
             if($current_page == $max_page) {
-                $paginate .= ' class="inactive"';
+                $paginate .= ' text-muted inactive"';
             } else {
-                $paginate .= ' onclick="$(\'filterPage\').value='.$next_page.';$(\'paginate\').submit();"';
+                $paginate .= '" onclick="$(\'filterPage\').value='.$next_page.';$(\'paginate\').submit();"';
             }
-            $paginate .= ' />';
+            $paginate .= '></i>';
 
         } else {