+++ include/inc_tmpl/admin.structlist.tmpl.php
@@ -51,9 +51,8 @@ echo "<td width=\"450\">";
 echo "<table class=\"table-borderless\"><tr>";
 echo '<td class="text-nowrap">';
 echo ($child_count) ? "<a href=\"phpwcms.php?do=admin&amp;p=6&amp;open=0:".(empty($_SESSION["structure"][0])?1:0)."\">" : "";
-echo "<img src=\"img/symbole/plus_".(($child_count) ? (empty($_SESSION["structure"][0]) ? "open" : "close") : "empty");
-echo ".gif\" width=\"15\" height=\"15\" border=\"0\" alt=\"\" />".(($child_count) ? "</a>" : "");
-echo '<img src="img/symbole/page_1.gif" width="11" height="15" alt="ID:0"';
+echo '<i class="far fa-'.(($child_count) ? (empty($_SESSION["structure"][0]) ? "plus-square" : "minus-square") : "square").' fa-fw text-muted" aria-hidden="true"></i>'.(($child_count) ? "</a>" : "");
+echo '<i class="far fa-file text-muted mx-1" title="ID:0" ';
 echo 'onmouseover="Tip(\'ID: <b>0</b><br>', $BL['be_alias'], ': ', html($indexpage["acat_alias"]);
 if(isset($struct_template[0]['template_name'])) {
 	echo '<br>', $BL['be_admin_struct_template'], ': ';